// apps/frontend/src/shared/components/Typography/Typography.types.ts
export type TypographyVariants =
  | "heading1"
  | "heading2"
  | "heading3"
  | "heading4"
  | "heading2El"
  | "heading3El"
  | "display1"
  | "display2"
  | "display3"
  | "display4"
  | "bodyLarge"
  | "bodyMedium"
  | "bodySmallL"
  | "bodySmallR"
  | "bodySmallM"
  | "bodySmallB"
  | "bodySmallH"
  | "captionL"
  | "captionR"
  | "captionM"
  | "captionB"
  | "disclaimerL"
  | "disclaimerR"
  | "disclaimerM"
  | "disclaimerB"
  | "category"
  | "appNav";

export type TypographyTags =
  | "address"
  | "div"
  | "figcaption"
  | "h1"
  | "h2"
  | "h3"
  | "h4"
  | "h5"
  | "h6"
  | "label"
  | "li"
  | "p"
  | "span";